requirements.txt
dbt-core==1.9.4
dbt-duckdb==1.9.3
dlt==1.10.0
docker==7.1.0
duckdb==1.2.2
gitdb==4.0.12
GitPython==3.1.41
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.0
gspread==6.2.0
numpy==2.2.5
pandas==2.2.3
prefect==3.4.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pyarrow==17.0.0
dlt[motherduck]